# api-contract.yaml
openapi: 3.1.0
info:
  title: Routa.js API Contract
  description: |
    Single source of truth for the Routa.js dual-backend API.
    Both the Next.js backend (src/app/api/) and the Rust backend (crates/routa-server/)
    MUST implement all endpoints defined here with compatible request/response shapes.
  version: 0.1.0
servers:
  - url: http://localhost:3000
    description: Next.js backend (dev)
  - url: http://localhost:3210
    description: Rust backend (desktop)

# ─────────────────────────────────────────────────────────
# Shared component schemas
# ─────────────────────────────────────────────────────────
components:
  schemas:
...
    VirtualFileEntry:
      type: object
      required: [path, name, isDirectory]
      properties:
        path:
          type: string
        name:
          type: string
        isDirectory:
          type: boolean
        size:
          type: integer
        children:
          type: array
          items:
            $ref: "#/components/schemas/VirtualFileEntry"

    # ── Models ──
...
paths:
...
  /api/clone:
    get:
      operationId: listClonedRepos
      summary: List cloned repositories
      responses:
        "200":
          description: Repo list
          content:
            application/json:
              schema:
                type: object
                properties:
                  repos:
                    type: array
                    items:
                      type: object
    post:
      operationId: cloneRepo
      summary: Clone a GitHub repository
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url:
                  type: string
      responses:
        "200":
          description: Cloned
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  path:
                    type: string
                  name:
                    type: string
                  branch:
                    type: string
                  branches:
                    type: array
                    items:
                      type: string
                  existed:
                    type: boolean
    patch:
      operationId: switchBranch
      summary: Switch branch on cloned repo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [repoPath, branch]
              properties:
                repoPath:
                  type: string
                branch:
                  type: string
      responses:
        "200":
          description: Switched
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  branch:
                    type: string
                  branches:
                    type: array
                    items:
                      type: string
...
  /api/github/import:
    post:
      operationId: importGitHubRepo
      summary: Import a GitHub repo as a virtual workspace (zipball download)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                owner:
                  type: string
                repo:
                  type: string
                ref:
                  type: string
                  description: Branch, tag, or SHA. Defaults to HEAD.
                url:
                  type: string
                  description: Shorthand — GitHub URL (parsed into owner/repo)
      responses:
        "200":
          description: Import successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  owner:
                    type: string
                  repo:
                    type: string
                  ref:
                    type: string
                  fileCount:
                    type: integer
                  extractedPath:
                    type: string
                  importedAt:
                    type: string
                    format: date-time
        "404":
          description: Repository not found
        "403":
          description: Rate limited or forbidden
        "413":
          description: Repository too large
  /api/github/issues:
    get:
      operationId: listGitHubIssues
      summary: List GitHub issues for a workspace codebase
      parameters:
        - name: workspaceId
          in: query
          required: true
          schema:
            type: string
        - name: codebaseId
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
            enum: [open, closed, all]
      responses:
        "200":
          description: GitHub issues
          content:
            application/json:
              schema:
                type: object
                properties:
                  repo:
                    type: string
                  codebase:
                    type: object
                    properties:
                      id:
                        type: string
                      label:
                        type: string
                  issues:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        number:
                          type: integer
                        title:
                          type: string
                        body:
                          type: string
                        url:
                          type: string
                        state:
                          type: string
                        labels:
                          type: array
                          items:
                            type: string
                        assignees:
                          type: array
                          items:
                            type: string
                        updatedAt:
                          type: string
        "400":
          description: Missing workspace or non-GitHub codebase
        "404":
          description: Workspace codebase not found
        "502":
          description: GitHub upstream request failed
  /api/github/tree:
    get:
      operationId: getGitHubTree
      summary: Get file tree for an imported GitHub repo
      parameters:
        - name: owner
          in: query
          required: true
          schema:
            type: string
        - name: repo
          in: query
          required: true
          schema:
            type: string
        - name: ref
          in: query
          schema:
            type: string
            default: HEAD
      responses:
        "200":
          description: File tree
          content:
            application/json:
              schema:
                type: object
                properties:
                  tree:
                    type: array
                    items:
                      $ref: "#/components/schemas/VirtualFileEntry"
                  fileCount:
                    type: integer
        "404":
          description: Workspace not imported
  /api/github/file:
    get:
      operationId: getGitHubFile
      summary: Read a file from an imported GitHub repo
      parameters:
        - name: owner
          in: query
          required: true
          schema:
            type: string
        - name: repo
          in: query
          required: true
          schema:
            type: string
        - name: path
          in: query
          required: true
          schema:
            type: string
        - name: ref
          in: query
          schema:
            type: string
            default: HEAD
      responses:
        "200":
          description: File content
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: string
                  path:
                    type: string
        "404":
          description: File or workspace not found
  /api/github/search:
    get:
      operationId: searchGitHubFiles
      summary: Search files in an imported GitHub repo
      parameters:
        - name: owner
          in: query
          required: true
          schema:
            type: string
        - name: repo
          in: query
          required: true
          schema:
            type: string
        - name: q
          in: query
          schema:
            type: string
        - name: ref
          in: query
          schema:
            type: string
            default: HEAD
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  files:
                    type: array
                    items:
                      type: object
                      properties:
                        path:
                          type: string
                        name:
                          type: string
                        score:
                          type: number
                  total:
                    type: integer
                  query:
                    type: string

  # ── Workspace sub-resources ──
...
  /api/webhooks/github:
    get:
      operationId: listGithubWebhooks
      summary: List registered GitHub webhooks
      responses:
        "200":
          description: Webhook list
    post:
      operationId: handleGithubWebhook
      summary: Handle an incoming GitHub webhook event
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Processed
...
  /api/github/pr-comment:
    post:
      operationId: postGithubPrComment
      summary: Post a comment on a GitHub pull request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [owner, repo, prNumber, body]
              properties:
                owner:
                  type: string
                repo:
                  type: string
                prNumber:
                  type: integer
                body:
                  type: string
      responses:
        "200":
          description: Comment posted

  # ── A2UI Dashboard ──
...